Fixed date being of type Double

+++ src/main/java/org/yorksolutions/teamobjbackend/entities/Product.java
@@ -47,7 +47,7 @@ public class Product
     @ElementCollection
     List<String> categories;
     @JsonProperty
-    Double startDate;
+    Long startDate;
     @JsonProperty
     Double defaultPrice;
     @JsonProperty
